requirements.txt
# Core Inference & Orchestration
vllm>=0.6.0
ray[default]>=2.9.0
langgraph>=0.2.0
langchain>=0.1.0
langchain-community>=0.0.20

# API & Communication
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
redis>=5.0.1
httpx>=0.26.0

# Model Management
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Memory & State Management
chromadb>=0.4.22
numpy>=1.26.0
psutil>=5.9.8

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=24.1.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tenacity>=8.2.3
asyncio>=3.4.3
aioredis>=2.0.1

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.1.14
mypy>=1.8.0

# ROCm Support (AMD GPU)
# Install separately: pip install torch torchvision --index-url https://download.pytorch.org/whl/rocm5.7
